Rename activity_interrupt_type.h to activity_interrupt.h
So that the file's name is the same as the enum (class).
