+++ crawl-ref/source/MSVC/crawl.vcxproj
@@ -829,7 +829,7 @@ perl.exe "util/gen-cflg.pl" compflag.h "&lt;UNKNOWN&gt;" "&lt;UNKNOWN&gt;"
     <ClInclude Include="..\ac-type.h" />
     <ClInclude Include="..\acquire.h" />
     <ClInclude Include="..\act-iter.h" />
-    <ClInclude Include="..\activity-interrupt-type.h" />
+    <ClInclude Include="..\activity-interrupt.h" />
     <ClInclude Include="..\actor.h" />
     <ClInclude Include="..\adjust.h" />
     <ClInclude Include="..\AppHdr.h" />
