+++ crawl-ref/source/MSVC/crawl.vcxproj.filters
@@ -954,7 +954,7 @@
     <ClInclude Include="..\act-iter.h">
       <Filter>h</Filter>
     </ClInclude>
-    <ClInclude Include="..\activity-interrupt-type.h">
+    <ClInclude Include="..\activity-interrupt.h">
       <Filter>h</Filter>
     </ClInclude>
     <ClInclude Include="..\actor.h">
