+++ crawl-ref/source/activity-interrupt.h
